// src/interfaces/Address.ts
export interface Address {
  country: string;
  city: string;
  province: string;
  postalCode: string;
  streetNumber: string;
  streetName: string;
  unit?: string;
  phoneNumber?: string;
  extension?: string;
  buildingType?: string;
  isGroundFloor?: string;
  elevatorAccess?: string;
  address?: string;
}